Build 7 - UI improvements
Add a new panel (PlayWorkoutController) to play the workout instead of putting everything in the first workout panel.
Add a new animated image sequence to show that we have started even if the song has not yet started playing.
Add skip button.
Improve the first visible tab by adding color gradient view that matches the first image.
Improve the display of the workout bars by adding a gradient and some shading.
Add countdown voice over when the song is finishing.